requirements.txt
mistralai==1.1.0
pdf2image==1.17.0
Pillow==11.0.0
pydantic==2.9.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
colorama~=0.4.6